requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision

albumentations
opencv-python

einops
tqdm

omegaconf
hydra-core